+++ NEWS.md
@@ -1,3 +1,12 @@
+peacesciencer 0.3.0 (in development)
+---------------------------------------------------------------------
+
+- Add `atop_alliance` (directed dyad-year) data
+- Add `add_atop_alliance()` function
+- Add `add_minimum_distance()` function
+- File compression for data files (again)
+- Rename variables in `cow_alliance` to distinguish from ATOP equivalents.
+
 
 peacesciencer 0.2.0
 ---------------------------------------------------------------------
@@ -7,7 +16,7 @@ peacesciencer 0.2.0
 - Add Correlates of War intergovernmental organizations data.
 - Add function for adding Correlates of War intergovernmental organizations data.
 - Add Correlates of War trade data (state-year and dyad-year).
-- Add minimum distance data and `add_minimum_distance()` function
+- Add minimum distance data
 - Add citations data frame for assistance in properly citing things.
 - Add `create_statedays()` function.